listening-comp/backend/requirements.txt
chromadb
streamlit
google-cloud-aiplatform
google-cloud-texttospeech
google-cloud-speech
google-generativeai
python-dotenv
youtube_transcript_api
requests